Don't markDirty() when inserting into a full slot
When inserting into a full slot, the inventory is not actually modified
but markDirty() is still called. This in turn sends a block update to
all neighbours. When a diamond chest is full, attempting to insert into
it will send 100+ block updates (one for each slot). This can cause mild
performance issues.